data processing

+++ burn_raster_driver.py
@@ -14,14 +14,16 @@ import utm
 
 
 #directory = 'R:\\LSDTopoTools\\Topographic_projects\\full_himalaya\\'
-directory = '/exports/csce/datastore/geos/users/s1134744/LSDTopoTools/Topographic_projects/full_himalaya_5000/'
+directory = '/exports/csce/datastore/geos/users/s1134744/LSDTopoTools/Topographic_projects/0_05_concavity/'
 
 #data_source = '/exports/csce/datastore/geos/users/s1134744/LSDTopoTools/Topographic_projects/TRMM_data/full_data/non_monsoon_trans.tif' 
 
 #data_source = 'R:\\exhumation\\0_2.tif'#source of the rasterized exhumation data
 data_source_exhumation = '/exports/csce/datastore/geos/users/s1134744/exhumation/0_2.tif'
 data_source_glaciated = '/exports/csce/datastore/geos/users/s1134744/LSDTopoTools/Topographic_projects/shapefiles/glims_ice/ice_100.bil'
+data_source_cosmo = '/exports/csce/datastore/geos/groups/LSDTopoData/Himalayan_Ksn_Concavity/cosmo_data/basin_shapefiles/outputDS.tif'
 data_source_distance = '/exports/csce/datastore/geos/users/s1134744/LSDTopoTools/Topographic_projects/shapefiles/simplified_distance.tif'
+data_source_distance_km = '/exports/csce/datastore/geos/users/s1134744/LSDTopoTools/Topographic_projects/shapefiles/distance_km.tif'
 data_source_distance_along = '/exports/csce/datastore/geos/users/s1134744/LSDTopoTools/Topographic_projects/shapefiles/euc_allocation_6.tif'
 data_source_strain = '/exports/csce/datastore/geos/users/s1134744/LSDTopoTools/Topographic_projects/strain/second_invariant.tif'
 
@@ -78,7 +80,7 @@ def getGlaciers(full_directory,summary_directory,fname,write_name,SRTM90=False):
   
   if os.path.isfile(full_directory+fname+'_glaciated.bil'):
     try:
-      os.remove(full_directory+fname+'_glaciated.bil')
+      os.remove(full_directory+fname+'_glaciated.bil')  
       os.remove(full_directory+fname+'_glaciated.hdr')
     except:
       print("Error removing raster files. Forcing exit to prevent unseen errors")
@@ -110,14 +112,16 @@ def getRaster(full_directory,summary_directory,fname,write_name,raster_source,ra
   
          
   #clipping raster, reprojecting, and setting resolution 
-  res_90 = "gdalwarp -ot Float64 -of ENVI -tr 90 90 -s_srs 'EPSG:4326' -t_srs"+" '"+ds+"' -te %s %s %s %s %s %s%s_%s.bil" %(LL[0],LL[1],UR[0],UR[1],raster_source,full_directory,fname,raster_name)
-  res_30 = "gdalwarp -ot Float64 -of ENVI -tr 30 30 -s_srs 'EPSG:4326' -t_srs"+" '"+ds+"' -te %s %s %s %s %s %s%s_%s.bil" %(LL[0],LL[1],UR[0],UR[1],raster_source,full_directory,fname,raster_name)    
+  #setting nodata values to ensure preservation
+  #changing from filename to writename
+  res_90 = "gdalwarp -ot Float64 -of ENVI -tr 90 90 -srcnodata '-9999' -dstnodata '-9999' -s_srs 'EPSG:4326' -t_srs"+" '"+ds+"' -te %s %s %s %s %s %s%s_%s.bil" %(LL[0],LL[1],UR[0],UR[1],raster_source,full_directory,write_name,raster_name)
+  res_30 = "gdalwarp -ot Float64 -of ENVI -tr 30 30 -srcnodata '-9999' -dstnodata '-9999' -s_srs 'EPSG:4326' -t_srs"+" '"+ds+"' -te %s %s %s %s %s %s%s_%s.bil" %(LL[0],LL[1],UR[0],UR[1],raster_source,full_directory,write_name,raster_name)    
   
   #before rasterising, check if raster is already present in directory, remove if it is
-  if os.path.isfile(full_directory+fname+'_'+raster_name+'.bil'):
+  if os.path.isfile(full_directory+write_name+'_'+raster_name+'.bil'):
     try:
-      os.remove(full_directory+fname+'_'+raster_name+'.bil')
-      os.remove(full_directory+fname+'_'+raster_name+'.hdr')
+      os.remove(full_directory+write_name+'_'+raster_name+'.bil')
+      os.remove(full_directory+write_name+'_'+raster_name+'.hdr')
     except:
       print("Error removing raster files. Forcing exit to prevent unseen errors")
       sys.exit()
@@ -165,14 +169,23 @@ for name in name_list:
             write_name = str(row[1])+str(row[5])+'_'+str(max_basin)
       
             #column_header = "digitized"
-            secondary_column_header = "Null"
+            #secondary_column_header = "tectonic_zone"
+            #logic to rename source DEM files/make sure they exist
+            
+            if not os.path.isfile(full_target+write_name+'.bil'):
+            
+                if not os.path.isfile(full_target+fname+'.bil'):
+                
+                    try:
+                        shutil.copy2(summary_target+fname+'.bil',full_target+fname+'.bil')
+                        shutil.copy2(summary_target+fname+'.hdr',full_target+fname+'.hdr')                        
+                    except:
+                        print("cannot find the dem for %s"%(fname))                        
+                        
+                os.rename(full_target+fname+'.bil',full_target+write_name+'.bil')
+                os.rename(full_target+fname+'.hdr',full_target+write_name+'.hdr')
+            
 
-            if not os.path.isfile(full_target+fname+'.bil'):
-                try:
-                    shutil.copy2(summary_target+fname+'.bil',full_target+fname+'.bil')
-                    shutil.copy2(summary_target+fname+'.hdr',full_target+fname+'.hdr')
-                except:
-                    print("cannot find the dem for %s"%(fname))
           
 
     
@@ -190,14 +203,15 @@ for name in name_list:
             #getRaster(full_target,summary_target,fname,write_name,data_source_distance,raster_name='distance')            
             #getRaster(full_target,summary_target,fname,write_name,data_source_distance_along,raster_name='distance_along')
             
-            getRaster(full_target,summary_target,fname,write_name,data_source_strain,raster_name='strain_ezz')            
+            getRaster(full_target,summary_target,fname,write_name,data_source_cosmo,raster_name='cosmo')
+            getRaster(full_target,summary_target,fname,write_name,data_source_tectonics,raster_name='tectonics')            
             
             #getGlaciers(full_target,summary_target,fname,write_name)
             #print(full_target,summary_target,fname,write_name)
             #try:
             #getGeologyRaster(full_target,summary_target,fname,write_name)
             #  sys.exit()
-            #except:
+            #except:                                                                                                    
             #  print("error")
             #sys.exit()
             #put into m_n list loop if adding to all concavity ksn calculations
@@ -210,8 +224,8 @@ for name in name_list:
                 x = x.replace('.','_')
     
                 try:
-                    os.rename(full_target+fname+'.bil',full_target+fname+x+'.bil')
-                    os.rename(full_target+fname+'.hdr',full_target+fname+x+'.hdr')
+                    os.rename(full_target+write_name+'.bil',full_target+write_name+x+'.bil')
+                    os.rename(full_target+write_name+'.hdr',full_target+write_name+x+'.hdr')
     
                 except(OSError):
                     print("ERROR! Probably a missing tile. Location: %s%s"%(full_target,fname))        
@@ -225,22 +239,22 @@ for name in name_list:
                     file.write('read path: %s \n'%(full_target))
                     file.write('write path: %s \n'%(full_target))
       
-                    file.write('read fname: %s \n'%(fname+x))
-                    file.write('write fname: %s \n'%(fname+x))
+                    file.write('read fname: %s \n'%(write_name+x))
+                    file.write('write fname: %s \n'%(write_name+x))
       
                     #file.write('read fname: %s \n'%(write_name))
                     #file.write('write fname: %s \n'%(write_name))      
       
                     file.write("\n")
                     file.write('burn_raster_to_csv: True \n')
-                    file.write('burn_raster_prefix: %s \n'%(fname+'_strain_ezz'))
+                    file.write('burn_raster_prefix: %s \n'%(fname+'_cosmo'))
                     #file.write('burn_raster_prefix: %s \n'%(fname+'_glaciated'))
                     #file.write('burn_data_csv_column_header: tertiary_burned_data \n')
-                    file.write('burn_data_csv_column_header: second_inv \n')
+                    file.write('burn_data_csv_column_header: cosmo_EBE_MMKYR \n')
                     file.write("\n")
-                    file.write('secondary_burn_raster_to_csv: False \n')
-                    file.write('secondary_burn_raster_prefix: %s \n'%(fname+'_distance_along'))
-                    file.write('secondary_burn_data_csv_column_header: distance_along \n')
+                    file.write('secondary_burn_raster_to_csv: True \n')
+                    file.write('secondary_burn_raster_prefix: %s \n'%(fname+'_tectonics'))
+                    file.write('secondary_burn_data_csv_column_header: tectonic_zone \n')
                     file.close()
                         
                 #os.remove(full_target+fname+"_Chiculations.param")
